@@ -1,15 +1,17 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
+plugins {
+    id 'com.android.library'
+    id 'org.jetbrains.kotlin.android'
+}
 
 android {
-    compileSdkVersion 28
+    namespace 'com.tencent.qgame.animplayer'
+    compileSdk 34
 
     defaultConfig {
-        minSdkVersion 16
-        targetSdkVersion 28
-        versionCode 1
-        versionName "1.0"
+        minSdk 21
+        targetSdk 34
+
+        consumerProguardFiles "proguard-rules.pro"
     }
 
     buildTypes {
@@ -19,18 +21,21 @@ android {
         }
     }
 
-}
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
+    }
 
-dependencies {
-    implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version") {
-        exclude module: 'annotations'
+    kotlinOptions {
+        jvmTarget = '17'
     }
 }
 
-// jcenter 上传(这个要在底部)
-// 上传需要执行此任务 IDE -> gradle-> Tasks/publishing/bintrayUpload
-// apply from: file("publish.gradle")
-
+dependencies {
+    implementation("org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version") {
+        exclude module: 'annotations'
+    }
+}
 
 // maven central
 // 上传指令./gradlew uploadArchives
