@@ -34,8 +34,7 @@ import com.tencent.qgame.animplayer.mask.MaskConfig
 import com.tencent.qgame.animplayer.util.ALog
 import com.tencent.qgame.animplayer.util.IALog
 import com.tencent.qgame.animplayer.util.ScaleType
-import com.tencent.qgame.playerproj.R
-import kotlinx.android.synthetic.main.activity_anim_simple_demo.*
+import com.tencent.qgame.playerproj.databinding.ActivityAnimSimpleDemoBinding
 import java.io.File
 import java.nio.ByteBuffer
 import java.util.zip.Inflater
@@ -52,6 +51,8 @@ class AnimActiveDemoActivity : Activity(), IAnimListener {
     companion object {
         private const val TAG = "AnimSimpleDemoActivity"
     }
+
+    private lateinit var binding: ActivityAnimSimpleDemoBinding
 
     private val dir by lazy {
         // 存放在sdcard应用缓存文件中
@@ -83,7 +84,8 @@ class AnimActiveDemoActivity : Activity(), IAnimListener {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_anim_simple_demo)
+        binding = ActivityAnimSimpleDemoBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         // 文件加载完成后会调用init方法
         val files = Array(1) {
             videoInfo.fileName
@@ -101,7 +103,7 @@ class AnimActiveDemoActivity : Activity(), IAnimListener {
         // 初始化调试开关
         initTestView()
         // 获取动画view
-        animView = playerView
+        animView = binding.playerView
         // 居中（根据父布局按比例居中并全部显示s）
         animView.setScaleType(ScaleType.FIT_CENTER)
         // 启动过滤遮罩s
@@ -218,18 +220,18 @@ class AnimActiveDemoActivity : Activity(), IAnimListener {
 
 
     private fun initTestView() {
-        btnLayout.visibility = View.VISIBLE
+        binding.btnLayout.visibility = View.VISIBLE
         /**
          * 开始播放按钮
          */
-        btnPlay.setOnClickListener {
+        binding.btnPlay.setOnClickListener {
             updateTestMask()
             play(videoInfo)
         }
         /**
          * 结束视频按钮
          */
-        btnStop.setOnClickListener {
+        binding.btnStop.setOnClickListener {
             animView.stopPlay()
         }
     }
@@ -278,4 +280,3 @@ class AnimActiveDemoActivity : Activity(), IAnimListener {
         return dp * scale + 0.5f
     }
 }
-
