@@ -29,8 +29,7 @@ import com.tencent.qgame.animplayer.inter.IAnimListener
 import com.tencent.qgame.animplayer.util.ALog
 import com.tencent.qgame.animplayer.util.IALog
 import com.tencent.qgame.animplayer.util.ScaleType
-import com.tencent.qgame.playerproj.R
-import kotlinx.android.synthetic.main.activity_anim_simple_demo.*
+import com.tencent.qgame.playerproj.databinding.ActivityAnimSimpleDemoBinding
 import java.io.File
 
 /**
@@ -41,6 +40,8 @@ class AnimSimpleDemoActivity : Activity(), IAnimListener {
     companion object {
         private const val TAG = "AnimSimpleDemoActivity"
     }
+
+    private lateinit var binding: ActivityAnimSimpleDemoBinding
 
     private val dir by lazy {
         // 存放在sdcard应用缓存文件中
@@ -62,7 +63,8 @@ class AnimSimpleDemoActivity : Activity(), IAnimListener {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_anim_simple_demo)
+        binding = ActivityAnimSimpleDemoBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         // 文件加载完成后会调用init方法
         loadFile()
     }
@@ -73,7 +75,7 @@ class AnimSimpleDemoActivity : Activity(), IAnimListener {
         // 初始化调试开关
         initTestView()
         // 获取动画view
-        animView = playerView
+        animView = binding.playerView
         // 居中（根据父布局按比例居中并全部显示，默认fitXY）
         animView.setScaleType(ScaleType.FIT_CENTER)
         // 注册动画监听
@@ -180,17 +182,17 @@ class AnimSimpleDemoActivity : Activity(), IAnimListener {
 
 
     private fun initTestView() {
-        btnLayout.visibility = View.VISIBLE
+        binding.btnLayout.visibility = View.VISIBLE
         /**
          * 开始播放按钮
          */
-        btnPlay.setOnClickListener {
+        binding.btnPlay.setOnClickListener {
             play(videoInfo)
         }
         /**
          * 结束视频按钮
          */
-        btnStop.setOnClickListener {
+        binding.btnStop.setOnClickListener {
             animView.stopPlay()
         }
     }
@@ -212,4 +214,3 @@ class AnimSimpleDemoActivity : Activity(), IAnimListener {
         return dp * scale + 0.5f
     }
 }
-
