@@ -36,7 +36,7 @@ import com.tencent.qgame.animplayer.util.ALog
 import com.tencent.qgame.animplayer.util.IALog
 import com.tencent.qgame.animplayer.util.ScaleType
 import com.tencent.qgame.playerproj.R
-import kotlinx.android.synthetic.main.activity_anim_simple_demo.*
+import com.tencent.qgame.playerproj.databinding.ActivityAnimSimpleDemoBinding
 import java.io.File
 import java.util.*
 
@@ -50,6 +50,8 @@ class AnimVapxDemoActivity : Activity(), IAnimListener {
     companion object {
         private const val TAG = "AnimSimpleDemoActivity"
     }
+
+    private lateinit var binding: ActivityAnimSimpleDemoBinding
 
     private val dir by lazy {
         // 存放在sdcard应用缓存文件中
@@ -74,7 +76,8 @@ class AnimVapxDemoActivity : Activity(), IAnimListener {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_anim_simple_demo)
+        binding = ActivityAnimSimpleDemoBinding.inflate(layoutInflater)
+        setContentView(binding.root)
         // 文件加载完成后会调用init方法
         loadFile()
     }
@@ -85,7 +88,7 @@ class AnimVapxDemoActivity : Activity(), IAnimListener {
         // 初始化调试开关
         initTestView()
         // 获取动画view
-        animView = playerView
+        animView = binding.playerView
         // 居中（根据父布局按比例居中并裁剪）
         animView.setScaleType(ScaleType.CENTER_CROP)
         /**
@@ -252,17 +255,17 @@ class AnimVapxDemoActivity : Activity(), IAnimListener {
 
 
     private fun initTestView() {
-        btnLayout.visibility = View.VISIBLE
+        binding.btnLayout.visibility = View.VISIBLE
         /**
          * 开始播放按钮
          */
-        btnPlay.setOnClickListener {
+        binding.btnPlay.setOnClickListener {
             play(videoInfo)
         }
         /**
          * 结束视频按钮
          */
-        btnStop.setOnClickListener {
+        binding.btnStop.setOnClickListener {
             animView.stopPlay()
         }
     }
@@ -284,4 +287,3 @@ class AnimVapxDemoActivity : Activity(), IAnimListener {
         return dp * scale + 0.5f
     }
 }
-
