@@ -18,28 +18,31 @@ package com.tencent.qgame.playerproj
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
+import com.tencent.qgame.playerproj.databinding.ActivityMainBinding
 import com.tencent.qgame.playerproj.player.AnimActiveDemoActivity
 import com.tencent.qgame.playerproj.player.AnimSimpleDemoActivity
 import com.tencent.qgame.playerproj.player.AnimSpecialSizeDemoActivity
 import com.tencent.qgame.playerproj.player.AnimVapxDemoActivity
-import kotlinx.android.synthetic.main.activity_main.*
 
 
 class MainActivity : Activity(){
 
+    private lateinit var binding: ActivityMainBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        btn1.setOnClickListener {
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        binding.btn1.setOnClickListener {
             startActivity(Intent(this, AnimSimpleDemoActivity::class.java))
         }
-        btn2.setOnClickListener {
+        binding.btn2.setOnClickListener {
             startActivity(Intent(this, AnimVapxDemoActivity::class.java))
         }
-        btn3.setOnClickListener {
+        binding.btn3.setOnClickListener {
             startActivity(Intent(this, AnimActiveDemoActivity::class.java))
         }
-        btn4.setOnClickListener {
+        binding.btn4.setOnClickListener {
             startActivity(Intent(this, AnimSpecialSizeDemoActivity::class.java))
         }
     }
